+++ src/test/java/io/codemine/java/postgresql/codecs/MoneyCodecIT.java
@@ -2,8 +2,8 @@ package io.codemine.java.postgresql.codecs;
 
 import io.codemine.java.postgresql.CodecITBase;
 
-public class MoneyCodecIT extends CodecITBase<Long> {
+public class MoneyCodecIT extends CodecITBase<Money> {
   public MoneyCodecIT() {
-    super(Codec.MONEY, Long.class);
+    super(Codec.money(2), Money.class);
   }
 }