+++ src/test/java/io/codemine/java/postgresql/codecs/MoneyCodecTest.java
@@ -2,8 +2,8 @@ package io.codemine.java.postgresql.codecs;
 
 import io.codemine.java.postgresql.CodecTestBase;
 
-public class MoneyCodecTest extends CodecTestBase<Long> {
+public class MoneyCodecTest extends CodecTestBase<Money> {
   public MoneyCodecTest() {
-    super(Codec.MONEY);
+    super(Codec.money(2));
   }
 }